backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
bcrypt==4.1.1
python-dateutil==2.8.2
